Fixed up tests.
Tests were failing due to some ambiguities in the version of zookeeper
that we are deploying. Fixed them so that they work with my setup (juju
2.0, installing zookeeper from the charms store), and have some checks
that should make them work with other setups.

Also fixed a mistake, where the deploy test was calling self.unit, but
had set self.kafka.