Fix for issues with Zookeeper client_bind_addr option.
Since this value was set at config time, the config would be duplicated
across instances of the charm, instead of being unique to each machine,
as an ip address should be.

This, coupled with matching code change in the bigtop base layer, fixes
the issue, and also fixes an issue where Zookeeper would not restart if
you changed its interface after deploying it.